Fix error
FIx error: "Warning: React does not recognize the `showDropdown` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `showdropdown` instead. If you accidentally passed it from a parent component, remove it from the DOM element."

React didn't appreciate that I passed a prop to a styled component that represents an SVG I guess